Put new generic methods behind a go:build directive
