+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/RogueConsultingDev/grust
 
-go 1.27
+go 1.24.0
 
 require (
 	github.com/jaswdr/faker/v2 v2.9.1
