+++ safetypes/result-go1.27_test.go
@@ -1,0 +1,3 @@
+//go:build go1.27
+
+package st
